backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
redis[asyncio]==5.2.1
pydantic[email]==2.10.4
anthropic==0.44.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
